pyproject.toml
[project]
name = "learn-dataeng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andy Miles", email = "arrivingskillsai@outlook.com" }
]
requires-python = ">=3.13"
dependencies = ["kafka-python-ng>=2.2"]

[project.scripts]
learn-dataeng = "learn_dataeng:main"
send-kafka-data = "learn_dataeng.send_kafka_data:main"
receive-kafka-data = "learn_dataeng.receive_kafka_data:main"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"
